Refactor: 1 invocation = 1 output, Pandera contract, dispatch pattern
Architecture change: replace 3 monolithic compute scripts (each producing
multiple methods in one CSV) with a single dispatcher + 3 private modules.

  compute_divergence.py --method S1_MMD --output tab_div_S1_MMD.csv

- DivergenceSchema (Pandera, strict=True) validates every output at write
- 16 Make targets in divergence.mk, proper per-channel dependencies
- _divergence_semantic.py, _divergence_lexical.py, _divergence_citation.py
  contain pure functions, no argparse, no main
- plot_divergence.py updated to derive method name from filename
- 28 tests pass (modules, smoke for all 16 methods, schema validation)
- Old monolithic scripts and per-channel plot/test files removed
